+++ RELEASE_NOTES.rst
@@ -5,6 +5,8 @@ Release notes
 ----
 
 * ``openstack-doc-test``: Sort entries in index.html file.
+* ``diff_branches.py``: Add options containing DEPRECATED in their help
+  string to the deprecation list.
 
 0.27
 ----